refactoring: use a closure to wrap transactions
This introduces the db.Client.WithTransaction func, which is used for
all transactional work in Steve.

This function makes sure either Commit or Rollback is called after
Beginning a transaction, thereby avoiding cases when a forgotten
open transaction can block all other operations.

This commit also heavily refactors transaction.Client and associated
code in order to make it as minimal and as close to the wrapped
sql.Tx functions as possble, in order to reduce cognitive load
when working with this part of the codebase.

Signed-off-by: Silvio Moioli <silvio@moioli.net>