Generate SQL for searching in explicit arrays.
This happens when you look for a string in field X, and while X isn't indexed, X[1], X[2], etc. are.

Or we can treat the field X as a string of values "x1|x2|x3|x4" and see if it contains
the search string "x" in one of the or-bar parts (not supporting partial matches).