Revert part of steve PR 633
Specifically, recomment a test that was uncommented in this PR,
and revert the underlying code that made this test pass.  That
code change causes user-based integration tests to fail, so something
is not correct in that code.

This was actually unrelated to the reason for PR 633, to fix
a problem when sorting on a label that dropped objects that
didn't have that label.  That part remains untouched.