Do more selective schema-based resetting
Initially we need to process all the schemas - creating watchers
and informers for each one.

Later when a CRD is changed, we only want to change the GVKs
associated with that CRD. But for whatever reason we need to
process all the schemas through the watcher-blinking.  This
code is probably a candidate for optimization as well.